+++ mex/editor/edit/main.py
@@ -26,6 +26,26 @@ def edit_title() -> rx.Component:
     )
 
 
+def deactivate_all_switch() -> rx.Component:
+    """Render a switch to deactivate all primary source and values."""
+    return rx.hstack(
+        rx.spacer(),
+        "deactive all",
+        rx.switch(
+            checked=EditState.any_primary_source_or_editor_value_enabled,
+            on_change=EditState.disable_all_primary_source_and_editor_values,
+            disabled=rx.cond(
+                EditState.any_primary_source_or_editor_value_enabled,
+                False,  # noqa: FBT003
+                True,  # noqa: FBT003
+            ),
+            color_scheme="jade",
+            custom_attrs={"data-testid": "deactivate-all-switch"},
+        ),
+        style={"width": "100%"},
+    )
+
+
 def index() -> rx.Component:
     """Return the index for the edit component."""
     return page(
@@ -33,6 +53,7 @@ def index() -> rx.Component:
             rule_page_header(
                 edit_title(),
             ),
+            deactivate_all_switch(),
             rx.foreach(
                 RuleState.fields,
                 editor_field,