+++ mex/editor/edit/state.py
@@ -32,3 +32,25 @@ class EditState(RuleState):
             params.pop("saved")
             if event := self.push_url_params(params):
                 yield event
+
+    @rx.var
+    def any_primary_source_or_editor_value_enabled(self) -> bool:
+        """Determine if any primary source or editor value is enabled.
+
+        Returns:
+            bool: Return true if any primary source or editor value is enabled,
+            otherwise false.
+        """
+        return any(
+            ps.enabled or any(value.enabled for value in ps.editor_values)
+            for field in self.fields
+            for ps in field.primary_sources
+        )
+
+    def disable_all_primary_source_and_editor_values(self) -> None:
+        """Disable all primary source and editor values."""
+        for field in self.fields:
+            for ps in field.primary_sources:
+                ps.enabled = False
+                for value in ps.editor_values:
+                    value.enabled = False