+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- added a switch to toggle all primary sources and values in the edit page
+- added textareas for fields listed in mex/editor/models.yaml
 - Warning when the create or edit browser window/tab is closed or navigated away
 
 ### Changes